# Backend/backend/src/main/java/com/gestorpyme/dto/request/CreateReceivableRequest.java
package com.gestorpyme.dto.request;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;

import java.math.BigDecimal;
import java.time.LocalDate;

public record CreateReceivableRequest(
        @NotBlank String descricao,
        @NotNull @Positive BigDecimal valor,
        @NotNull LocalDate vencimento
) {}