# Backend/backend/src/main/java/com/gestorpyme/dto/request/LoginRequest.java
package com.gestorpyme.dto.request;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank @Email String email,
        @NotBlank String senha
) {}