# Backend/backend/src/main/java/com/gestorpyme/dto/response/ErrorResponse.java
package com.gestorpyme.dto.response;

import java.time.Instant;

public record ErrorResponse(
        Instant timestamp,
        int status,
        String error,
        String message,
        String path
) {}